Remove some ethereum stuff from the main roll
In case anyone ever reads the commit messages here... I now think
Ethereum has been a pretty negative thing for the world since 2014, and
while I'd rather do a post about that instead of deleting stuff, I'd
rather not have it be the case that this is the bulk of the content for
now. Posts remain available, but are now a bit harder to find.